+++ src/converse-controlbox.js
@@ -69,7 +69,7 @@ converse.plugins.add('converse-controlbox', {
      *
      * NB: These plugins need to have already been loaded via require.js.
      */
-    dependencies: ["converse-modal", "converse-chatboxes", "converse-rosterview", "converse-chatview"],
+    dependencies: ["converse-modal", "converse-chatboxes", "converse-chat", "converse-rosterview", "converse-chatview"],
 
     enabled (_converse) {
         return !_converse.singleton;
@@ -626,16 +626,25 @@ converse.plugins.add('converse-controlbox', {
              * @namespace _converse.api.controlbox
              * @memberOf _converse.api
              */
-            'controlbox': {
+            controlbox: {
                 /**
-                 * Retrieves the controlbox view.
-                 *
+                 * Opens the controlbox
+                 * @method _converse.api.controlbox.open
+                 * @returns { Promise<_converse.ControlBox> }
+                 */
+                async open () {
+                    await _converse.api.waitUntil('chatBoxesFetched');
+                    const model = await _converse.api.chatboxes.get('controlbox') ||
+                      _converse.api.chatboxes.create('controlbox', {}, _converse.Controlbox);
+                    model.trigger('show');
+                    return model;
+                },
+
+                /**
+                 * Returns the controlbox view.
                  * @method _converse.api.controlbox.get
-                 *
-                 * @example
-                 * const view = _converse.api.controlbox.get();
-                 *
-                 * @returns {Backbone.View} View representing the controlbox
+                 * @returns { Backbone.View } View representing the controlbox
+                 * @example const view = _converse.api.controlbox.get();
                  */
                 get () {
                     return _converse.chatboxviews.get('controlbox');
