+++ src/headless/converse-core.js
@@ -78,10 +78,12 @@ const CORE_PLUGINS = [
     'converse-bosh',
     'converse-caps',
     'converse-chatboxes',
+    'converse-chat',
     'converse-disco',
     'converse-emoji',
     'converse-mam',
     'converse-muc',
+    'converse-headlines',
     'converse-ping',
     'converse-pubsub',
     'converse-roster',
